requirements.txt
torch~=1.13.0
wandb~=0.13.4
pytest~=7.2.0
pytorch-lightning
jsonargparse[signatures]
hydra-core
pre-commit
omegaconf~=2.2.3
tueplots
numpy~=1.23.4
matplotlib~=3.6.1
PyYAML~=6.0
torchmetrics~=0.10.2
setuptools~=65.5.0
torchvision~=0.14.0
lightning-bolts~=0.5.0